added playerservice & modified float64 to uint for non-negative values

+++ backend/api/models/saloon.go
@@ -5,6 +5,6 @@ type Saloon struct {
 	Name        string  `json:"name" gorm:"unique"`
 	TalbleCount uint64  `json:"table_count" gorm:"column:tables_count"`
 	PlayerCount uint64  `json:"player_count" gorm:"column:players_count"`
-	MinBet      float64 `json:"min_bet"`
-	MaxBet      float64 `json:"max_bet"`
+	MinBet      uint64  `json:"min_bet"`
+	MaxBet      uint64  `json:"max_bet"`
 }
